+++ R/Spectra-functions.R
@@ -77,7 +77,8 @@ NULL
 #' @param ... optional additional arguments to `FUN`.
 #'
 #' @param f `factor` or `vector` that can be coerced to one defining how the
-#'     data should be split for parallel processing.
+#'     data should be split for parallel processing. Set to `NULL` to disable
+#'     parallel processing.
 #'
 #' @param columns `character` defining the columns that should be returned. This
 #'     will be passed to the backend's `peaksData` function.
@@ -96,24 +97,32 @@ NULL
                         f = dataStorage(object),
                         columns = c("mz", "intensity"), BPPARAM = bpparam()) {
     len <- length(object)
+    lf <- length(f)
     if (!len)
         return(list())
-    if (length(f) != len)
+    if (lf && lf != len)
         stop("Length of 'f' has to match 'length(object)' (", len, ")")
     if (!is.factor(f))
         f <- factor(f)
     pqueue <- object@processingQueue
     if (!is.null(FUN))
         pqueue <- c(pqueue, ProcessingStep(FUN, ARGS = list(...)))
-    if (length(levels(f)) > 1 || length(pqueue)) {
-        res <- bplapply(split(object@backend, f), function(z, queue, svars) {
+    ll <- length(levels(f))
+    if (length(pqueue) || ll > 1) {
+        .local <- function(z, queue, svars) {
             if (length(svars))
                 spd <- as.data.frame(spectraData(z, columns = svars))
             else spd <- NULL
             .apply_processing_queue(peaksData(z, columns = columns), spd,
                                     queue = queue)
-        }, queue = pqueue, svars = spectraVariables, BPPARAM = BPPARAM)
-        unsplit(res, f = f, drop = TRUE)
+        }
+        if (ll > 1) {
+            res <- bplapply(split(object@backend, f), .local,
+                            queue = pqueue, svars = spectraVariables,
+                            BPPARAM = BPPARAM)
+            unsplit(res, f = f, drop = TRUE)
+        } else
+            .local(object@backend, pqueue, spectraVariables)
     } else peaksData(object@backend, columns = columns)
 }
 
@@ -168,6 +177,7 @@ applyProcessing <- function(object, f = dataStorage(object),
     if (isReadOnly(object@backend))
         stop(class(object@backend), " is read-only. 'applyProcessing' works ",
              "only with backends that support writing data.")
+    BPPARAM <- backendBpparam(object@backend, BPPARAM)
     if (!is.factor(f))
         f <- factor(f, levels = unique(f))
     if (length(f) != length(object))
@@ -280,14 +290,27 @@ applyProcessing <- function(object, f = dataStorage(object),
 .compare_spectra_chunk <- function(x, y = NULL, MAPFUN = joinPeaks,
                                    tolerance = 0, ppm = 20,
                                    FUN = ndotproduct,
-                                   chunkSize = 10000, ...) {
+                                   chunkSize = 10000, ...,
+                                   BPPARAM = bpparam()) {
     nx <- length(x)
     ny <- length(y)
+    bppx <- backendBpparam(x@backend, BPPARAM)
+    if (is(bppx, "SerialParam"))
+        fx <- NULL
+    else fx <- factor(dataStorage(x))
+    pqvx <- .processingQueueVariables(x)
+    bppy <- backendBpparam(y@backend, BPPARAM)
+    if (is(bppy, "SerialParam"))
+        fy <- NULL
+    else fy <- factor(dataStorage(y))
+    pqvy <- .processingQueueVariables(y)
     if (nx <= chunkSize && ny <= chunkSize) {
-        mat <- .peaks_compare(.peaksapply(x), .peaksapply(y), MAPFUN = MAPFUN,
-                              tolerance = tolerance, ppm = ppm,
-                              FUN = FUN, xPrecursorMz = precursorMz(x),
-                              yPrecursorMz = precursorMz(y), ...)
+        mat <- .peaks_compare(
+            .peaksapply(x, f = fx, spectraVariables = pqvx, BPPARAM = bppx),
+            .peaksapply(y, f = fy, spectraVariables = pqvy, BPPARAM = bppy),
+            MAPFUN = MAPFUN, tolerance = tolerance, ppm = ppm,
+            FUN = FUN, xPrecursorMz = precursorMz(x),
+            yPrecursorMz = precursorMz(y), ...)
         dimnames(mat) <- list(spectraNames(x), spectraNames(y))
         return(mat)
     }
@@ -304,11 +327,19 @@ applyProcessing <- function(object, f = dataStorage(object),
                   dimnames = list(spectraNames(x), spectraNames(y)))
     for (x_chunk in x_chunks) {
         x_buff <- x[x_chunk]
-        x_peaks <- .peaksapply(x_buff)
+        if (length(fx))
+            fxc <- factor(dataStorage(x_buff))
+        else fxc <- NULL
+        x_peaks <- .peaksapply(x_buff, f = fxc, spectraVariables = pqvx,
+                               BPPARAM = bppx)
         for (y_chunk in y_chunks) {
             y_buff <- y[y_chunk]
+            if (length(fy))
+                fyc <- factor(dataStorage(y_buff))
+            else fyc <- NULL
             mat[x_chunk, y_chunk] <- .peaks_compare(
-                x_peaks, .peaksapply(y_buff),
+                x_peaks, .peaksapply(y_buff, f = fyc, spectraVariables = pqvy,
+                                     BPPARAM = bppy),
                 MAPFUN = MAPFUN, tolerance = tolerance, ppm = ppm,
                 FUN = FUN, xPrecursorMz = precursorMz(x_buff),
                 yPrecursorMz = precursorMz(y_buff), ...)
@@ -341,15 +372,19 @@ applyProcessing <- function(object, f = dataStorage(object),
     m <- matrix(NA_real_, nrow = nx, ncol = nx,
                   dimnames = list(spectraNames(x), spectraNames(x)))
 
+    sv <- .processingQueueVariables(x)
     cb <- which(lower.tri(m, diag = TRUE), arr.ind = TRUE)
     pmz <- precursorMz(x)
+    bpp <- SerialParam()
     for (i in seq_len(nrow(cb))) {
         cur <- cb[i, 2L]
         if (i == 1L || cb[i - 1L, 2L] != cur) {
-            py <- px <- peaksData(x[cur])[[1L]]
+            py <- px <- .peaksapply(x[cur], spectraVariables = sv, f = NULL,
+                                    BPPARAM = bpp)[[1L]]
             pmzx <- pmzy <- pmz[cur]
         } else {
-            py <- peaksData(x[cb[i, 1L]])[[1L]]
+            py <- .peaksapply(x[cb[i, 1L]], spectraVariables = sv, f = NULL,
+                              BPPARAM = bpp)[[1L]]
             pmzy <- pmz[cb[i, 1L]]
         }
         map <- MAPFUN(px, py, tolerance = tolerance, ppm = ppm,
@@ -409,8 +444,9 @@ applyProcessing <- function(object, f = dataStorage(object),
     x_new <- x[match(levels(f), f)]
     if (isReadOnly(x_new@backend))
         x_new <- setBackend(x_new, MsBackendMemory())
+    bpp <- SerialParam()
     peaksData(x_new@backend) <- lapply(
-        split(.peaksapply(x, BPPARAM = SerialParam()), f = f), FUN = FUN, ...)
+        split(.peaksapply(x, f = NULL, BPPARAM = bpp), f = f), FUN = FUN, ...)
     x_new@processingQueue <- list()
     validObject(x_new)
     x_new
@@ -464,6 +500,7 @@ combineSpectra <- function(x, f = x$dataStorage, p = x$dataStorage,
         f <- factor(f, levels = unique(f))
     if (!is.factor(p))
         p <- factor(p, levels = unique(p))
+    BPPARAM <- backendBpparam(x@backend, BPPARAM)
     if (length(f) != length(x) || length(p) != length(x))
         stop("length of 'f' and 'p' have to match length of 'x'")
     if (isReadOnly(x@backend))
@@ -611,6 +648,7 @@ estimatePrecursorIntensity <- function(x, ppm = 10, tolerance = 0,
     if (is.factor(f))
         f <- as.character(f)
     f <- factor(f, levels = unique(f))
+    BPPARAM <- backendBpparam(x@backend, BPPARAM)
     unlist(bplapply(split(x, f), FUN = .estimate_precursor_intensity, ppm = ppm,
                     tolerance = tolerance, method = method, msLevel = msLevel.,
                     BPPARAM = BPPARAM), use.names = FALSE)
@@ -642,7 +680,7 @@ estimatePrecursorIntensity <- function(x, ppm = 10, tolerance = 0,
     pmi <- rep(NA_real_, length(pmz))
     idx_ms2 <- which(!is.na(pmz) & msLevel(x) == msLevel)
     x_ms1 <- filterMsLevel(x, msLevel = msLevel - 1L)
-    pks <- .peaksapply(x_ms1)
+    pks <- .peaksapply(x_ms1, f = NULL, BPPARAM = SerialParam())
     for (i in idx_ms2) {
         rt_i <- rtime(x[i])
         before_idx <- which(rtime(x_ms1) < rt_i)